feat(lxc/machine-learning): use home manager only for machine-learning host
Unable to get the nvidia drivers installed in nixos lxc and without the
kernel modules. I will stick to the current ubuntu setup that works with
home manager managing the ml user configs. It's an okay compromise and
still allows for declaring and updating AI/ML apps/services via config